all in one seed file.
