+++ homeworks/week11/hw1/public/index.html
@@ -1,15 +1,23 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <title>Document</title>
+  <title>user / comment</title>
 </head>
+
 <body>
-  <form action="/submit" method="post">
-    <textarea name="comment" id="" cols="30" rows="10"></textarea>
-    <button type="submit">submit</button>
+  <form action="/addComment" method="post">
+    username: <input type="text" name="user" id="user"><br>
+    password: <input type="password" name="password" id="password"><br>
+    nickname: <input type="text" name="nickname" id="nickname"><br>
+    comment: <textarea name="comment" id="comment" cols="10" rows="10"></textarea><br>
+    sub-comment: <textarea name="sub_comment" id="sub_comment" cols="5" rows="10"></textarea><br>
+    <button type="submit">create a comment</button>
   </form>
+
 </body>
+
 </html>
