+++ homeworks/week11/hw1/seedData/data.js
@@ -1,4 +1,234 @@
-'use strict';
+
+const usernames = [
+  "Scott and Zelda Fitzgerald",
+  "Jean Cocteau",
+  "Brick top",
+  "Hemingway",
+  "Mark Twain",
+  "Sertride Stein",
+  "Belmonte",
+  "Pablo",
+  "Adriana",
+  "Amedeo",
+  "Modigliani",
+  "Cole Porter",
+  "Define",
+  "Archibald MacLeish",
+  "Dejuan Barnes",
+  "Dal",
+  "Señor Bunuel",
+  "Man Ray",
+  "Tom Stearns Eliot",
+  "Matisse",
+  "Paul Gaugin",
+  "Degas",
+  "Lautrec",
+  "Richard",
+];
+
+const nicknames = [
+  "Bad Mr. Frosty",
+  "Kraken",
+  "Boomer",
+  "Lumberjack",
+  "Boomerang",
+  "Mammoth",
+  "Boss",
+  "Mastadon",
+  "Budweiser",
+  "Master",
+  "Bullseye",
+  "Meatball",
+  "Buster",
+  "Mooch",
+  "Butch",
+  "Mr. President",
+  "Buzz",
+  "Outlaw",
+  "Canine",
+  "Ratman",
+  "Captian RedBeard",
+  "Renegade",
+  "Champ",
+  "Sabertooth",
+  "Coma",
+  "Scratch",
+  "Crusher",
+  "Sentinel",
+  "Diesel",
+  "Speed",
+  "Doctor",
+  "Spike",
+  "Dreads",
+  "Subwoofer",
+  "Frankenstein",
+  "Thunderbird",
+  "Froggy",
+  "Tornado",
+  "General",
+  "Troubleshoot",
+  "Godzilla",
+  "Vice",
+  "Hammerhead",
+  "Viper",
+  "Handy Man",
+  "Wasp",
+  "Hound Dog",
+  "Wizard",
+  "Indominus",
+  "Zodiac",
+  "King Kong",
+  "Rocky",
+  "Caesar",
+  "Godzilla",
+  "Robin Hood",
+  "Joker",
+  "Kraken",
+  "Butch Cassidy",
+  "Goldfinger",
+  "Frankenstein",
+  "Sundance Kid",
+  "Sherlock",
+  "Terminator",
+  "Superman",
+  "Watson",
+  "Mad Max",
+  "Tarzan",
+  "Grinch",
+  "Tin Man",
+  "Rooster",
+  "Popeye",
+  "Scarecrow",
+  "Moses",
+  "Bullwinkle",
+  "Chewbacca",
+  "Jesus",
+  "Daffy",
+  "Han Solo",
+  "Zorro",
+  "Porky",
+  "Captain Redbeard",
+  "Batman",
+  "Mr. Magoo",
+  "Blackbeard",
+  "Lincoln",
+  "Jetson",
+  "HAL",
+  "Hannibal",
+  "Panther",
+  "Wizard",
+  "Darth Vader",
+  "Gumby",
+  "Zodiac",
+  "Alien",
+  "Underdog",
+  "V-Mort",
+  "The Shark",
+  "Sylvester",
+  "C-Brown",
+  "Martian",
+  "Space Ghost",
+  "Finch",
+  "Dracula",
+  "Felix",
+  "Indiana",
+  "Kevorkian",
+  "Jungle Man",
+  "Agent 99",
+  "Mad Dog",
+  "Alias",
+  "Moonshine",
+  "Blueberry",
+  "Mustang",
+  "Black Lotus",
+  "Neptune",
+  "Black Widow",
+  "Number Six",
+  "Bonbon",
+  "Opaline",
+  "Butterfly",
+  "Princess Peach",
+  "Champagne",
+  "Radar",
+  "Chardonnay",
+  "Sgt. Calhoun",
+  "Claws",
+  "Shadow",
+  "Copycat",
+  "Sphinx",
+  "Dahlia",
+  "Subzero",
+  "Dollface",
+  "Sugar Plum",
+  "Dragonfly",
+  "Trinity",
+  "Eclipse",
+  "Twister",
+  "Epiphany",
+  "Twizzler",
+  "The Evil Queen",
+  "Uhura",
+  "Firefly",
+  "Venus",
+  "Frostbite",
+  "Vicious",
+  "Hurricane",
+  "Vivi",
+  "Jessica Rabbit",
+  "Wind",
+  "Jetta",
+  "Winger",
+  "Katniss",
+  "Winter",
+  "Lady Luck",
+  "Wonder Woman",
+  "Lunar",
+  "Zelda",
+  "Cinderella",
+  "Turanga",
+  "Alias",
+  "Ella",
+  "Killer Frost",
+  "Number Six",
+  "Katniss",
+  "Batgirl",
+  "Princess Peach",
+  "Scarlet",
+  "Betty Boop",
+  "Trinity",
+  "Granger",
+  "Goldilocks",
+  "Venom",
+  "Princess",
+  "Reno",
+  "Poison Ivy",
+  "The Bride",
+  "Jo Jo",
+  "Superwoman",
+  "Poppins",
+  "Sabriel",
+  "Zelda",
+  "Dorothy",
+  "Ezzie",
+  "Lady Luck",
+  "Juno",
+  "Sphinx",
+  "Wonder Woman",
+  "Wicked",
+  "Isis",
+  "Uhura",
+  "Mrs. Robinson",
+  "Aphrodite",
+  "Jessica Rabbit",
+  "Baby Doll",
+  "Artemis",
+  "The Evil Queen",
+  "Zena",
+  "Athena",
+  "Snow White",
+  "Rapunzel",
+  "Cupid"
+];
 
 const comments = [
   "人類要黴菌只是為了它的酵素。黴菌可能需要空氣，但是酵素不需要。"
@@ -27,39 +257,7 @@ const comments = [
   , "拓美這兩批壽司飯總重約9公斤，他此時肩負大約350萬個微小神明的命運。"
   , "蓬萊米比較短而黏，壽司米屬於這種。"
 ]
-//
-
-let userIds = []
-let commentIds = []
-
-for(let i=1; i<25; i++){
-  userIds.push(i)
-}
-for(let i=1; i<21; i++){
-  commentIds.push(i)
-}
-let subPosts = []
-
-for (let i = 0; i < 50; i++) {
-  let content = comments[Math.floor(Math.random() * comments.length)]
-  let user_id = userIds[Math.floor(Math.random() * userIds.length)]
-  let comment_id = commentIds[Math.floor(Math.random() * commentIds.length)]
-  let post = {
-    content,
-    user_id,
-    comment_id,
-    createdAt: new Date(),
-    updatedAt: new Date(),
-  }
-  subPosts.push(post)
-}
-  // console.log(post);
 
 module.exports = {
-  up: (queryInterface, Sequelize) => {
-    return queryInterface.bulkInsert('subComments', subPosts);
-  },
-
-  down: (queryInterface, Sequelize) => {
-  }
-};
+  usernames, nicknames, comments
+}
